+++ src/br/com/dotcom/entidade/Planeta.java
@@ -2,16 +2,16 @@ package br.com.dotcom.entidade;
 
 public class Planeta {
 
-	private long id;
+	private Integer id;
 	private String nome;
 	private String clima;
 	private String terreno;
 	private String qtdAparicao;
 	
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
 	public String getNome() {
